stk_usecases:  Removed gcc-4.7.2 compiler warnings.
(From Sierra)

+++ packages/stk/stk_usecases/app/UseCase_7.cpp
@@ -627,14 +627,14 @@ void use_case_7_generate_mesh(
             const int * const local_node_id = & elem_conn[ j * 8 ] ;
 
             const stk::mesh::EntityId node_id[8] = {
-              node_map[ local_node_id[0] - 1 ] ,
-              node_map[ local_node_id[1] - 1 ] ,
-              node_map[ local_node_id[2] - 1 ] ,
-              node_map[ local_node_id[3] - 1 ] ,
-              node_map[ local_node_id[4] - 1 ] ,
-              node_map[ local_node_id[5] - 1 ] ,
-              node_map[ local_node_id[6] - 1 ] ,
-              node_map[ local_node_id[7] - 1 ]
+              static_cast<uint64_t>(node_map[ local_node_id[0] - 1 ]),
+              static_cast<uint64_t>(node_map[ local_node_id[1] - 1 ]),
+              static_cast<uint64_t>(node_map[ local_node_id[2] - 1 ]),
+              static_cast<uint64_t>(node_map[ local_node_id[3] - 1 ]),
+              static_cast<uint64_t>(node_map[ local_node_id[4] - 1 ]),
+              static_cast<uint64_t>(node_map[ local_node_id[5] - 1 ]),
+              static_cast<uint64_t>(node_map[ local_node_id[6] - 1 ]),
+              static_cast<uint64_t>(node_map[ local_node_id[7] - 1 ])
             };
 
             const stk::mesh::EntityId elem_id = elem_map[ j ];
@@ -651,10 +651,10 @@ void use_case_7_generate_mesh(
             const int * const local_node_id = & elem_conn[ j * 4 ] ;
 
             const stk::mesh::EntityId node_id[4] = {
-              node_map[ local_node_id[0] - 1 ] ,
-              node_map[ local_node_id[1] - 1 ] ,
-              node_map[ local_node_id[2] - 1 ] ,
-              node_map[ local_node_id[3] - 1 ]
+              static_cast<uint64_t>(node_map[ local_node_id[0] - 1 ]),
+              static_cast<uint64_t>(node_map[ local_node_id[1] - 1 ]),
+              static_cast<uint64_t>(node_map[ local_node_id[2] - 1 ]),
+              static_cast<uint64_t>(node_map[ local_node_id[3] - 1 ])
             };
 
             const stk::mesh::EntityId elem_id = elem_map[ j ];
